refactored overall
